[PUCDR-1982] Fix crash when camera error occurs
Camera2Controller.java
   mDeviceStateCallback
      onError(@NonNull CameraDevice cameraDevice, int error)

The error may occur when cameraserver died or other low level issue.
DashCam will crash before becasue it happens IllegalStateException which want to close a camera that already closed (when error).

Solution is ignore the exception because we still want to try closing, then fail is OK.